+++ academiapoo/src/main/java/br/com/academia/poo/repository/ClienteRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import br.com.academia.poo.model.Cliente;
 
 
-public interface ClienteRepository extends CrudRepository<Cliente, Long> {
+public interface ClienteRepository extends CrudRepository<Cliente, Integer> {
 	List<Cliente> findByName(String name);
 }
