+++ academiapoo/src/main/java/br/com/academia/poo/repository/EquipamentoRepository.java
@@ -7,6 +7,6 @@ import org.springframework.data.repository.CrudRepository;
 import br.com.academia.poo.model.Equipamento;
 
 
-public interface EquipamentoRepository extends CrudRepository<Equipamento, Long> {
+public interface EquipamentoRepository extends CrudRepository<Equipamento, Integer> {
 	List<Equipamento> findByName(String name);
 }
